refactor: densify growth-stage relocation pin under 400 lines (#4310)
Move relocation Game fixtures into growth_stage_planner_test_support and
split AC7/AC14 case bodies into sibling *_cases.dart modules so the pin
contract drops from 411 to 10 physical lines.

Refs #4310
